pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "referral-engine"
version = "0.1.0"
description = "Pluggable, async-native multi-level referral distribution engine"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Franklin", email = "franklin0alex@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.setuptools]
packages = ["referral_engine", "referral_engine.adapters"]

[project.optional-dependencies]
db = [
    "asyncpg>=0.29.0",
    "pyyaml>=6.0.1",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic-settings>=2.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]